backend/requirements.txt
acachecontrol==0.3.6
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
boolean.py==5.0
CacheControl==0.14.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.4
cryptography==46.0.3
cyclonedx-python-lib==11.6.0
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.3.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
fastapi-cli==0.0.20
fastapi-cloud-cli==0.11.0
fastar==0.8.0
filelock==3.24.3
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
license-expression==30.4.4
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
okta-jwt-verifier==0.3.0
okta_jwt==1.3.5
packageurl-python==0.17.6
packaging==26.0
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.10.0
platformdirs==4.9.2
pluggy==1.6.0
prometheus_client==0.24.1
propcache==0.4.1
py-serializable==2.1.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
pyparsing==3.3.2
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
requests==2.32.5
retry2==0.9.5
rich==14.2.0
rich-toolkit==0.17.1
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.49.0
setuptools==80.10.1
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sortedcontainers==2.4.0
starlette==0.50.0
tomli==2.4.0
tomli_w==1.2.0
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
watchfiles==1.1.1
websockets==16.0
wrapt==2.1.0
yarl==1.22.0
